Round W8A8 probe reference to BF16 output
Both the grouped op and the per-layer fallback emit BF16, while the
reference stayed FP32, so err_fp32 included the output quantization
delta (~64 on this probe's ~16k-magnitude outputs).  Round the stacked
reference to BF16 before comparing, keeping the scale BF16-rounded as
well.

+++ benchmarks/probe_grouped_matmul.py
@@ -209,7 +209,9 @@ def _ref_w8a8(x: torch.Tensor, w8_ints, scale_2d) -> torch.Tensor:
             * scale.unsqueeze(0)
             * s_l
         )
-    return torch.stack(refs, dim=0)
+    # The ops output BF16; round the reference so err_fp32 excludes output
+    # quantization (on ~16k-magnitude outputs that alone is ~64).
+    return torch.stack(refs, dim=0).to(torch.bfloat16).float()
 
 
 def _make_variants(
